add an http client and basic sockets to test http
This commit adds everything we need for http. It reads in from the
command line and generates the http request and sends it off to the socket.
There is some temporary socket code just for testing purposes. The next step
is to create the tcp/ip headers and manage reliable delivery